+++ packages/flutter_gemma/lib/core/model_management/model_specs.dart
@@ -16,6 +16,7 @@ import 'package:flutter_gemma/core/model_management/constants/preferences_keys.d
 part 'types/model_spec.dart';
 part 'types/inference_model_spec.dart';
 part 'types/embedding_model_spec.dart';
+part 'types/stt_model_spec.dart';
 part 'types/storage_info.dart';
 part 'exceptions/model_exceptions.dart';